Move new interface processing to form clean method
